Remove improperly mapped glyphs from cmap
The Lohit Kannada and Telugu fonts contained cmap mappings for several
characters that were not valid. These mappings were conflicting with
valid glyphs from another font. This patch simply removes the mappings,
as follows:

1. Removed cmap mapping for U+2212 MINUS SIGN from both fonts that
pointed to a tofu glyph.
2. Removed cmap mapping for U+FFBC..FFF2 from the Telugu font that
pointed to tofu glyphs.

This patch is part of the fix for bug 7467609.

Change-Id: Id18850a296abc6667907166da0d7e072fb7b3f45